% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_collection.R
\name{add_collection_format}
\alias{add_collection_format}
\title{Download and install an image collection format from a URL}
\usage{
add_collection_format(url, name = NULL)
}
\arguments{
\item{url}{URL pointing to the collection format JSON file}

\item{name}{optional name used to refer to the collection format}
}
\description{
Download and install an image collection format from a URL
}
\details{
By default, the collection format name will be derived from the basename of the URL.
}
\examples{
\donttest{
add_collection_format(
   "https://raw.githubusercontent.com/appelmar/gdalcubes/dev/formats/Sentinel1_IW_GRD.json")
}
}
