% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{apply_pixel.array}
\alias{apply_pixel.array}
\title{Apply a function over pixels in a four-dimensional (band, time, y, x) array}
\usage{
\method{apply_pixel}{array}(x, FUN, ...)
}
\arguments{
\item{x}{four-dimensional input bands with dimension order band, time, y, x}

\item{FUN}{function which receives a vector of band values in a one-dimensional array}

\item{...}{further arguments passed to FUN}
}
\description{
Apply a function over pixels in a four-dimensional (band, time, y, x) array
}
\details{
FUN is expected to produce a numeric vector (or scalar) where elements are interpreted as new bands in the result
}
\note{
This is a helper function that uses the same dimension ordering as gdalcubes streaming. It can be used to simplify 
the application of R functions e.g. over time series in a data cube.
}
\examples{
\dontrun{
load(system.file("extdata","sample_chunk.Rdata", package="gdalcubes"))
y = apply_pixel(sample_chunk, function(x) {
 ndvi <- (x[8]-x[4])/(x[8]+x[4])
 return(c(ndvi=(x[8]-x[4])/(x[8]+x[4]), nir=x[8]))
})}
}
