% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{as_ncdf}
\alias{as_ncdf}
\title{"cube_view<-" <-function(x,value) {
  stopifnot(is.cube(x))
  stopifnot(is.cube_view(value))
  if (!is.image_collection_cube)
    stop("x is no image_collection_cube, updating the data cube view is currently only implemented for image_collection_cube")
  libgdalcubes_update_cube_view(x,value)
  return(x)
}
Materialize a data cube as a NetCDF file}
\usage{
as_ncdf(x, fname = tempfile(pattern = "as_ncdf_", fileext = ".nc"))
}
\arguments{
\item{x}{a data cube proxy object (class cube)}

\item{fname}{output file name}
}
\description{
This function will read chunks of a data cube and write them to a single NetCDF file.
}
