% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{bands}
\alias{bands}
\title{Query data cube properties}
\usage{
bands(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
A data.frame with rows representing the bands and columns representing properties of a band (name, type, scale, offset, unit)
}
\description{
Query data cube properties
}
