% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_apply.R
\name{chunk_apply}
\alias{chunk_apply}
\title{Apply a function on chunks of a data cube 
d}
\usage{
chunk_apply(cube, f)
}
\arguments{
\item{cube}{Source data cube}

\item{f}{R function to apply over all chunks}
}
\value{
A proxy data cube object
}
\description{
Apply a function on chunks of a data cube 
d
}
\details{
This function internally creates gdalcubes stream data cube, which streams
data of a chunk to stdin of a new R process. For reading data the function f typically 
calls \code{x <- read_chunk_as_array()} which then results in a 4 dimensional (band, time, y, x) array.
Similarly \code{write_stream_from_array(x)} will write a result array as a chunk in the resulting data cube.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{
f <- function() {
  x = read_chunk_as_array()
  out <- reduce_time(x, function(x) {
    y = x[1,]
    mean(y, na.rm=T)})
  write_chunk_from_array(out)
}
xstrm <- chunk_apply(XXX, f)
}
}
