% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Create a data cube from an image collection}
\usage{
cube(image_collection, view, chunking = c(16, 256, 256))
}
\arguments{
\item{image_collection}{Source image collection as from \code{image_collection} or \code{create_image_collection}}

\item{view}{A data cube view defining the shape (spatiotemporal extent, resolution, and spatial reference), if missing, a default overview is used}

\item{chunking}{Vector of length 3 defining the size of data cube chunks in the order time, y, x.}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which loads data from a given image collection
}
\details{
The following steps will be performed when the data cube is requested to read data of a chunk:

 1. Filter images from the input collection that intersect with the spatiotemporal extent of the chunk
 2. For all resulting images, apply gdalwarp to reproject / warp the image to the target SRS and size to a GDAL MEM dataset
 3. Read the resulting data to the chunk buffer and if pixels already contain non NAN values, apply an aggregation method (as defined in the view)
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
 L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                        ".TIF", recursive = TRUE, full.names = TRUE)
 v = cube_view(l=388941.2, r=766552.4, b=4345299, t=4744931, 
         proj="EPSG:32618",
         nx = 497, ny=526, t0="2018-01", t1="2018-12", dt="P1M")
 L8.col = create_image_collection(L8_files, "L8_L1TP") 
 cube(L8.col, v)
 
}
