% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{dim.cube}
\alias{dim.cube}
\title{Query data cube properties}
\usage{
\method{dim}{cube}(x)
}
\arguments{
\item{x}{a data cube proxy object (class cube)}
}
\value{
size of a data cube (number of cells) as integer vector in the order t, y, x
}
\description{
Query data cube properties
}
\seealso{
\code{\link{size}}
}
