% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{dim.gcbs_cube}
\alias{dim.gcbs_cube}
\title{Query data cube properties}
\usage{
\method{dim}{gcbs_cube}(x)
}
\arguments{
\item{x}{a data cube proxy object (class gcbs_cube)}
}
\value{
Size of a data cube (number of cells) as integer vector in the order t, y, x
}
\description{
Query data cube properties
}
\seealso{
gcbs_size
}
