% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{dimension_values}
\alias{dimension_values}
\title{Query coordinate values for all dimensions of a data cube}
\usage{
dimension_values(obj, datetime_unit = NULL)
}
\arguments{
\item{obj}{a data cube proxy (class cube), or a data cube view object}

\item{datetime_unit}{unit used to format values in the datetime dimension, one of "Y", "m", "d", "H", "M", "S", defaults to the unit of the cube.}
}
\value{
list with elements t,y,x
}
\description{
Dimension values give the coordinates along the spatial and temporal axes of a data cube.
}
