% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.default_chunk_size}
\alias{.default_chunk_size}
\title{Calculate a default chunk size based on the cube size and currently used number of threads}
\usage{
.default_chunk_size(nt, ny, nx)
}
\description{
Calculate a default chunk size based on the cube size and currently used number of threads
}
\examples{
.default_chunk_size(12, 1000, 1000)
}
