% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_time.R
\name{fill_time}
\alias{fill_time}
\title{Fill NA data cube pixels by time series interpolation}
\usage{
fill_time(cube, method = "near")
}
\arguments{
\item{cube}{source data cube}

\item{method}{interpolation method, can be "near" or "linear"}
}
\value{
a proxy data cube object
}
\description{
Create a proxy data cube, which fills NA pixels of a data cube by nearest neighbor or linear time series interpolation.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
