% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geom.R
\name{filter_geom}
\alias{filter_geom}
\title{Filter data cube pixels by a polygon}
\usage{
filter_geom(cube, geom, srs = NULL)
}
\arguments{
\item{cube}{source data cube}

\item{geom}{either a WKT string, or an sfg object}

\item{srs}{string identifier of spatial reference system ()}
}
\value{
a proxy data cube object
}
\description{
Create a proxy data cube, which filters pixels by a spatial (multi)polygon For all pixels whose center is within the polygon, the original
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
TODO
}
