% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pixel.R
\name{gcbs_apply_pixel}
\alias{gcbs_apply_pixel}
\title{Apply arithmetic expressions over all pixels of a data cube}
\usage{
gcbs_apply_pixel(cube, expr, names = NULL)
}
\arguments{
\item{cube}{Source data cube}

\item{expr}{character vector with one or more arithmetic expressions (see Details)}

\item{names}{optional character vector with the same length as expr to specify band names for the output cube}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which applies arithmetics expressions over all pixels of a data cube. Expressions may access band values by their name.
}
\details{
gdalcubes uses the \href{https://github.com/ArashPartow/exprtk}{exprtk library} to evaluate expressions in C++, you can look at the library examples 
to see what kind of expressions you can execute.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{
gcbs_apply_pixel(XXXX, expr=c("(B08-B04)/(B08+B04)"))
}
}
