% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_collection.R
\name{gcbs_collection_formats}
\alias{gcbs_collection_formats}
\title{List predefined image collection formats}
\usage{
gcbs_collection_formats(print = TRUE)
}
\arguments{
\item{print}{Should available formats and their descriptions be printed nicely?}
}
\value{
A data.frame with columns name and description where the former describes the unique identifier that can be used in \code{gcbs_create_image_collection} and the
latter gives a brief description of the format.
}
\description{
gdalcubes comes with some predefined collection formats e.g. to scan Sentinel 2 data. This function lists available formats and a brief description.
}
\details{
Image collection formats define how individul files / GDALDatasets relate to a collection, i.e., 
which bands they contain, to which image they belong, and how the derive date/time.
They are described as a set of regular expressions in a JSON file and used by gdalcubes to extract this information 
from the paths and/or filenames.
}
