% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_dimensions}
\alias{gcbs_dimensions}
\title{Query data cube properties}
\usage{
gcbs_dimensions(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
Dimension information as a data.frame, where each row represents a dimension and columns represent properties such as dimension boundaries, names, and chunk size
}
\description{
Query data cube properties
}
