% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_eval}
\alias{gcbs_eval}
\title{"gcbs_view<-" <-function(x,value) {
  stopifnot(is.gcbs_cube(x))
  stopifnot(is.gcbs_view(value))
  if (!is.gcbs_image_collection_cube)
    stop("x is no image_collection_cube, updating the data cube view is currently only implemented for image_collection_cube")
  libgdalcubes_update_cube_view(x,value)
  return(x)
}
Materialize a data cube as a NetCDF file}
\usage{
gcbs_eval(x, fname = tempfile(pattern = "gcbs_eval_", fileext = ".nc"))
}
\arguments{
\item{x}{a data cube proxy object (class gcbs_cube)}

\item{fname}{output file name}
}
\description{
This function will read chunks of a data cube and write them to a single NetCDF file.
}
