% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicate.R
\name{gcbs_filter_predicate}
\alias{gcbs_filter_predicate}
\title{Filter data cube pixels by a user-defined predicate on band values}
\usage{
gcbs_filter_predicate(cube, pred)
}
\arguments{
\item{cube}{Source data cube}

\item{pred}{predicate to be evaluated over all pixels}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which evaluates a predicate over all pixels of a data cube. For all pixels which fulfill the predicate, the original
band values are returned. Other pixels are simply filled with NANs. The predicate may access band values by their name.
}
\details{
gdalcubes uses the \href{https://github.com/ArashPartow/exprtk}{exprtk library} to evaluate expressions in C++, you can look at the library examples 
to see what kind of expressions you can execute.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{
gcbs_filter_predicate(XXXX, expr=c("(B08-B04)/(B08+B04)"))
}
}
