% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_get_projection}
\alias{gcbs_get_projection}
\title{Query data cube properties}
\usage{
gcbs_get_projection(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
The spatial reference system expressed as a string readable by GDAL
}
\description{
Query data cube properties
}
