% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_graph}
\alias{gcbs_graph}
\title{Query data cube properties}
\usage{
gcbs_graph(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
A JSON string representing a graph (currently a tree) that can be used to create the same
chain of gdalcubes operations.
}
\description{
gdalcubes uses a graph (currently a tree) to represent / serialize data cube chains. This function gives a JSON
representation, which will be communicated to gdalcubes_server instances to create identical cube instances 
remotely.
}
