% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_collection.R
\name{gcbs_image_collection}
\alias{gcbs_image_collection}
\title{Load an existing image collection from a file}
\usage{
gcbs_image_collection(path)
}
\arguments{
\item{path}{Path to an existing image collection file}
}
\value{
An image collection proxy object, which can be used to create a data cube using \code{gcbs_cube}
}
\description{
This function will load an image collection from an SQLite file. These files
index / reference existing imagery. To create a collection from files on disk,
use \code{gcbs_create_collection}.
}
\examples{
L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                       ".TIF", recursive = TRUE, full.names = TRUE)
gcbs_create_image_collection(L8_files, "L8_L1TP", out_file = "L8.db", quiet=TRUE)
gcbs_image_collection("L8.db")
}
