% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_nbands}
\alias{gcbs_nbands}
\title{Query data cube properties}
\usage{
gcbs_nbands(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
Number of bands
}
\description{
Query data cube properties
}
