% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_nt}
\alias{gcbs_nt}
\title{Query data cube properties}
\usage{
gcbs_nt(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
Number of pixels in the time dimension
}
\description{
Query data cube properties
}
