% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_nx}
\alias{gcbs_nx}
\title{Query data cube properties}
\usage{
gcbs_nx(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
Number of pixels in the x dimension
}
\description{
Query data cube properties
}
