% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{gcbs_ny}
\alias{gcbs_ny}
\title{Query data cube properties}
\usage{
gcbs_ny(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class gcbs_cube)}
}
\value{
Number of pixels in the y dimension
}
\description{
Query data cube properties
}
