% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{gcbs_read_stream_as_array}
\alias{gcbs_read_stream_as_array}
\title{Read chunk data of a cube from stdin}
\usage{
gcbs_read_stream_as_array(with.dimnames = TRUE)
}
\arguments{
\item{with.dimnames}{if TRUE, the resulting array will contain dimnames with coordinates, datetime, and band names}
}
\value{
four-dimensional array
}
\description{
gdalcubes stream cubes write chunk data to stdin of external processes (such as R)
This function can be used to read the data from stdin as four-dimensional R array with dimensions
(band, time, y, x).
}
\note{
This function only works in R sessions started from gdalcubes streaming
}
