% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{gcbs_reduce_time}
\alias{gcbs_reduce_time}
\title{Reduce a data cube over the time dimension}
\usage{
gcbs_reduce_time(cube, expr, ...)
}
\arguments{
\item{cube}{Source data cube}

\item{expr}{Either a single string, or a vector of strings defining which reducers wlil be applied over which bands of the input cube}

\item{...}{Optional additional expressions (if expr is not a vector)}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which applies one ore more reducer functions over selected bands of pixel time series of a data cube
}
\details{
Notice that expressions have a very simple format: the reducer is followed by the name of a band in parantheses. You cannot add
more complex functions or arguments.

Possible reducers currently are "min", "max", "sum", "prod", "count", "mean", "median", "var", and "sd".
}
\note{
Implemented reducers will ignore any NAN values (as na.rm=TRUE does)

This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{
gcbs_reduce_time(XXXX, "min(B08)", "max(B08)", median("B04"))}
}
