% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bands.R
\name{gcbs_select_bands}
\alias{gcbs_select_bands}
\title{Select bands of a data cube}
\usage{
gcbs_select_bands(cube, bands)
}
\arguments{
\item{cube}{Input data cube}

\item{bands}{character vector with band names}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which selects specific bands of a data cube. The resulting cube
will drop any other bands.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.

For performance reasons, gcbs_select_bands should always be called directly on a image collection cube created with gcbs_cube and 
drop all unneded bands.This allows to reduce GDAL RasterIO and warp operations.
}
\examples{
\dontrun{gcbs_select_bands()}
}
