% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{gcbs_server_status}
\alias{gcbs_server_status}
\title{Status report of gdalcubes_server instances}
\usage{
gcbs_server_status()
}
\value{
a data.frame with detailed status information where each row corresponds to one gdalcubes_server process
}
\description{
Summarizes the status of all gdalcubes_server processes that have been started in the current R session.
}
