% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gcbs_set_threads}
\alias{gcbs_set_threads}
\title{Set the number of threads for parallel data cube processing}
\usage{
gcbs_set_threads(n = 1)
}
\arguments{
\item{n}{number of threads}
}
\value{
nothing / invisible
}
\description{
Data cubes can be processed in parallel where one thread processes one chunk at a time. Setting more threads
than the number of chunks of a cube thus has no effect and will not furhter reduce computation times.
}
