% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{gcbs_stop_server}
\alias{gcbs_stop_server}
\title{Stop all gdalcubes_server processes}
\usage{
gcbs_stop_server()
}
\description{
Stops all gdalcubes_server processes that have been started within the current R session.
}
