% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gcbs_version}
\alias{gcbs_version}
\title{Query gdalcubes version information}
\usage{
gcbs_version()
}
\value{
List with gdalcubes library version information
}
\description{
Query gdalcubes version information
}
