% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{gcbs_write_stream_from_array}
\alias{gcbs_write_stream_from_array}
\title{Write chunk data of a cube to stdout}
\usage{
gcbs_write_stream_from_array(v)
}
\arguments{
\item{v}{A four-dimensional array with dimensions band, time, y, and x}
}
\description{
gdalcubes stream cubes write chunk data to stdin of external processes (such as R) and 
reads result data from stdout. 
This function can be used to write an R array as a reuslt chunk to stdout.
}
\note{
This function only works in R sessions started from gdalcubes streaming
}
