% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_debug_output}
\alias{gdalcubes_debug_output}
\title{Enable or disable debug output from the gdalcubes C++ library}
\usage{
gdalcubes_debug_output(debug = TRUE)
}
\arguments{
\item{debug}{logical, TRUE if you want debug messages}
}
\description{
Enable or disable debug output from the gdalcubes C++ library
}
\examples{
gdalcubes_debug_output(TRUE)
gdalcubes_debug_output(FALSE)
}
