% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_gdal_has_geos}
\alias{gdalcubes_gdal_has_geos}
\title{Check if GDAL was built with GEOS}
\usage{
gdalcubes_gdal_has_geos()
}
\description{
Check if GDAL was built with GEOS
}
\examples{
gdalcubes_gdal_has_geos()
}
