% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_gdalversion}
\alias{gdalcubes_gdalversion}
\title{Get the GDAL version used by gdalcubes}
\usage{
gdalcubes_gdalversion()
}
\description{
Get the GDAL version used by gdalcubes
}
\examples{
gdalcubes_gdalversion()
}
