% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_set_gdal_config}
\alias{gdalcubes_set_gdal_config}
\title{Set GDAL config options}
\usage{
gdalcubes_set_gdal_config(key, value)
}
\arguments{
\item{key}{name of a GDAL config option to be set}

\item{value}{value}
}
\description{
Set GDAL config options
}
\details{
Details and a list of possible options can be found at 
\href{https://gdal.org/user/configoptions.html}{https://gdal.org/user/configoptions.html}.
}
\examples{
gdalcubes_set_gdal_config("GDAL_NUM_THREADS", "ALL_CPUS")
}
