% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_set_threads}
\alias{gdalcubes_set_threads}
\title{Set the number of threads for parallel data cube processing}
\usage{
gdalcubes_set_threads(n = 1)
}
\arguments{
\item{n}{number of threads}
}
\description{
Data cubes can be processed in parallel where one thread processes one chunk at a time. Setting more threads
than the number of chunks of a cube thus has no effect and will not further reduce computation times.
}
\note{
THIS FUNCTION IS DEPRECATED AND IS GOING TO BE REPLACED BY \code{\link{gdalcubes_options}}.
}
\examples{
gdalcubes_set_threads(1)
}
