% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{gdalcubes_stop_server}
\alias{gdalcubes_stop_server}
\title{Stop all gdalcubes_server processes}
\usage{
gdalcubes_stop_server()
}
\description{
Stops all gdalcubes_server processes that have been started within the current R session.
}
\examples{
gdalcubes_stop_server() 
}
