% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_use_cache}
\alias{gdalcubes_use_cache}
\title{Enable or disable caching of cubes.}
\usage{
gdalcubes_use_cache(enable = TRUE)
}
\arguments{
\item{enable}{logical, TRUE if you want to use the data cube cache}
}
\description{
Enable or disable caching of cubes.
}
\details{
Caching has no effect on disk or memory consumption, 
it simply tries to reuse existing temporary files where possible.
For example, changing only parameters to \code{plot} will not require
rerunning the full data cube operation chain.
}
\note{
THIS FUNCTION IS DEPRECATED AND IS GOING TO BE REPLACED BY \code{\link{gdalcubes_options}}.
}
\examples{
gdalcubes_use_cache(FALSE)
}
