% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_version}
\alias{gdalcubes_version}
\title{Query gdalcubes version information}
\usage{
gdalcubes_version()
}
\value{
List with gdalcubes library version information
}
\description{
Query gdalcubes version information
}
\examples{
gdalcubes_version()
}
