% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_bands.R
\name{join_bands}
\alias{join_bands}
\title{Join bands of two identically shaped data cubes}
\usage{
join_bands(A, B)
}
\arguments{
\item{A}{The first source data cube}

\item{B}{The second source data cube}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which joins the bands two identically shaped data cubes. The resulting cube
will have bands from both input cubes.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{join_bands(XXXX, YYYY)}
}
