% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{memsize}
\alias{memsize}
\title{Query data cube properties}
\usage{
memsize(obj, unit = "MiB")
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}

\item{unit}{Unit of data size, can be "B", "KB", "KiB", "MB", "MiB", "GB", "GiB", "TB", "TiB", "PB", "PiB"}
}
\value{
Total data size of data cube values expressed in the given unit
}
\description{
Query data cube properties
}
