% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{names.cube}
\alias{names.cube}
\title{Query data cube properties}
\usage{
\method{names}{cube}(x)
}
\arguments{
\item{x}{a data cube proxy object (class cube)}
}
\value{
Band names as character vector
}
\description{
Query data cube properties
}
