% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{names.gcbs_cube}
\alias{names.gcbs_cube}
\title{Query data cube properties}
\usage{
\method{names}{gcbs_cube}(x)
}
\arguments{
\item{x}{a data cube proxy object (class gcbs_cube)}
}
\value{
Band names as character vector
}
\description{
Query data cube properties
}
