% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{nbands}
\alias{nbands}
\title{Query data cube properties}
\usage{
nbands(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
Number of bands
}
\description{
Query data cube properties
}
