% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_cube.R
\name{ncdf_cube}
\alias{ncdf_cube}
\title{Read a data cube from an existing netCDF file}
\usage{
ncdf_cube(path, chunking = NULL, auto_unpack = TRUE)
}
\arguments{
\item{path}{path to an existing netCDF file}

\item{chunking}{custom chunk sizes to read form the netCDF file; defaults to using chunk sizes from the netCDF file}

\item{auto_unpack}{logical; automatically apply offset and scale when reading data values}
}
\value{
a proxy data cube object
}
\description{
Create a proxy data cube, which reads a data cube from a netCDF file that has been created using \code{\link{write_ncdf}}.
This function does not read cubes from arbitrary netCDF files and can be used e.g., to load intermediate results and/or 
plotting existing netCDF cubes on disk without doing the data cube creation from image collections.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
TODO


}
