% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{nt}
\alias{nt}
\title{Query data cube properties}
\usage{
nt(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
Number of pixels in the time dimension
}
\description{
Query data cube properties
}
