% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{nx}
\alias{nx}
\title{Query data cube properties}
\usage{
nx(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
Number of pixels in the x dimension
}
\description{
Query data cube properties
}
