% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{ny}
\alias{ny}
\title{Query data cube properties}
\usage{
ny(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
Number of pixels in the y dimension
}
\description{
Query data cube properties
}
