% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cube}
\alias{predict.cube}
\title{Predict values based on a trained model to all pixels of a data cube.}
\usage{
\method{predict}{cube}(object, model, ...)
}
\arguments{
\item{object}{a data cube proxy object (class cube)}

\item{model}{model used for prediction (e.g. from \code{caret})}

\item{...}{further arguments (not used)}
}
\description{
Predict values based on a trained model to all pixels of a data cube.
}
\details{
The predict method used will be automatically chosen based on the model class. It should support models from the \code{caret} 
package and simple models as from\code{lm}. However, it  currently works only for a single output variable and 
requires a \code{data.frame} input.
}
\examples{
# TODO

}
