% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{print.cube_view}
\alias{print.cube_view}
\title{Print data cube view information}
\usage{
\method{print}{cube_view}(x, ...)
}
\arguments{
\item{x}{Object of class "cube_view"}

\item{...}{Further arguments passed to the generic print function}
}
\description{
Prints information about a data cube view, including its dimensions, spatial reference, aggregation method, and resampling method.
}
\examples{
v = cube_view(extent=list(left=388941.2, right=766552.4, 
              bottom=4345299, top=4744931, t0="2018-01", t1="2018-12"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")
print(v)
}
