% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_collection.R
\name{print.image_collection}
\alias{print.image_collection}
\title{Print image collection information}
\usage{
\method{print}{image_collection}(x, ..., n = 6)
}
\arguments{
\item{x}{Object of class "image_collection"}

\item{...}{Further arguments passed to the generic print function}

\item{n}{Number of images for which details are printed}
}
\description{
Prints information about images in an image collection.
}
\examples{
L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                       ".TIF", recursive = TRUE, full.names = TRUE)
v = cube_view(extent=list(left=388941.2, right=766552.4, 
              bottom=4345299, top=4744931, t0="2018-01", t1="2018-12"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")
L8.col = create_image_collection(L8_files, "L8_L1TP") 
print(L8.col)
}
