% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_points.R
\name{query_points}
\alias{query_points}
\title{Query data cube values at irregular spatiotemporal points}
\usage{
query_points(x, px, py, pt, srs)
}
\arguments{
\item{x}{source data cube}

\item{px}{vector of x coordinates}

\item{py}{vector of y coordinates}

\item{pt}{vector of date/ time coordinates}

\item{srs}{spatial reference system string identifer (as GDAL understands)}
}
\value{
a data.frame with one row per point and one column per data cube band or variable
}
\description{
The function will overlay provided spatiotemporal points with the data cube and return all band values
of the cells per point, as a data.frame where rows correspoind to points and columns represent bands.
If needed, point coordinates are automativally converted to the SRS of the data cube.
}
\details{
TODO: what date/ time format?

TODO: what das GDAL unterstand?
}
\examples{
# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db")) 
}
TODO

}
