% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{read_chunk_as_array}
\alias{read_chunk_as_array}
\title{Read chunk data of a data cube from stdin or a file}
\usage{
read_chunk_as_array(with.dimnames = TRUE)
}
\arguments{
\item{with.dimnames}{if TRUE, the resulting array will contain dimnames with coordinates, datetime, and band names}
}
\value{
four-dimensional array
}
\description{
This function can be used within function passed to \code{\link{chunk_apply}} in order to read a data cube chunk as a four-dimensional R array.
It works only for R processes, which have been started from the gdalcubes C++ library. 
The resulting array has dimensions band, time, y, x (in this order).
}
\note{
Call this function ONLY from a function passed to \code{\link{chunk_apply}}.

This function only works in R sessions started from gdalcubes streaming.
}
