% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\title{Reduce a data cube over the time dimension}
\usage{
reduce(cube, reducer = c("mean", "median", "min", "max"))
}
\arguments{
\item{cube}{Source data cube}

\item{reducer}{Reducer function, currently "min", "max", "median", "mean", "count", "sd", "var", or "sum"}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which applies a reducer function over pixel time series of a data cube
}
\note{
Implemented reducers will ignore any NAN values (as na.rm=TRUE does)

This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.

This function is deprecated and will be replaced by the mor flexible reduce_time.
}
\examples{
 L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                        ".TIF", recursive = TRUE, full.names = TRUE)
 v = cube_view(l=388941.2, r=766552.4, b=4345299, t=4744931, 
         proj="EPSG:32618",
         nx = 497, ny=526, t0="2018-01", t1="2018-12", dt="P1M")
 L8.col = create_image_collection(L8_files, "L8_L1TP") 
 L8.cube = cube(L8.col, v) 
 L8.rgb = select_bands(L8.cube, c("B02", "B03", "B04"))
 L8.rgb.median = reduce(L8.rgb, "median")  
 L8.rgb.median
 plot(L8.rgb.median, rgb=3:1, zlim=c(4000,12000))
}
