% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{size}
\alias{size}
\title{Query data cube properties}
\usage{
size(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
Size of a data cube (number of cells) as integer vector in the order t, y, x
}
\description{
Query data cube properties
}
\seealso{
dim.cube
}
