% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{srs}
\alias{srs}
\title{Query data cube properties}
\usage{
srs(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
The spatial reference system expressed as a string readable by GDAL
}
\description{
Query data cube properties
}
