% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{stac_image_collection}
\alias{stac_image_collection}
\title{Create an image collection from a STAC feature collection}
\usage{
stac_image_collection(
  s,
  out_file = tempfile(fileext = ".sqlite"),
  asset_names = NULL,
  asset_regex = NULL,
  url_fun = function(x) {     paste0("/vsicurl/", x) }
)
}
\arguments{
\item{s}{STAC feature collection}

\item{asset_names}{character vector with names of assets (e.g., bands) to be used, other assets will be ignored. By default (NULL), all asset names with "eo:bands" attributes will be used}

\item{asset_regex}{length 1 character defining a regular expression asset names must match to be considered}

\item{url_fun}{optional function to modify URLs of assets, e.g, to add /vsicurl/ to URLS (the default)}
}
\description{
This function is experimental.
}
\note{
Currently, bbox results are expected to be WGS84 coordinates, even if bbox-crs is given in the STAC response.
}
