% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_time.cube}
\alias{window_time.cube}
\title{Apply a moving window function over the time dimension of a data cube}
\usage{
\method{window_time}{cube}(x, window, expr, ...)
}
\arguments{
\item{x}{Source data cube}

\item{window}{integer vector with two elements defining the size of the window before and after a cell, the total size ofof the window is window[1] + 1 + window[2]}

\item{expr}{Either a single string, or a vector of strings defining which reducers wlil be applied over which bands of the input cube}

\item{...}{Optional additional expressions (if expr is not a vector)}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which applies one ore more moving window reducer functions over selected bands of pixel time series of a data cube
}
\details{
Notice that expressions have a very simple format: the reducer is followed by the name of a band in parantheses. You cannot add
more complex functions or arguments.

Possible reducers currently are "min", "max", "sum", "prod", "count", "mean", "median".
}
\note{
Implemented reducers will ignore any NAN values (as na.rm=TRUE does)

This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                       ".TIF", recursive = TRUE, full.names = TRUE)
v = cube_view(l=388941.2, r=766552.4, b=4345299, t=4744931, 
              proj="EPSG:32618",
              nx = 497, ny=526, t0="2018-01", t1="2018-12", dt="P1M")
L8.col = create_image_collection(L8_files, "L8_L1TP") 
L8.cube = cube(L8.col, v) 
L8.nir = select_bands(L8.cube, c("B08"))
L8.nir.min = window_time(L8.rgb, c(2,2), "min(B02)")  
L8.nir.min
plot(L8.nir.min, zlim=c(4000,12000), key.pos=1, t=c(1,4,7))

}
