% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_time}
\alias{window_time}
\title{Apply a moving window operation over time}
\usage{
window_time(x, ...)
}
\arguments{
\item{object}{to be reduced}
}
\value{
return value and type depend on the class of x
}
\description{
This generic function applies a reducer function over a moving window over the time dimension of a data cube, an R array, or other classes if implemented.
}
