% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{write_chunk_from_array}
\alias{write_chunk_from_array}
\title{Write chunk data of a cube to stdout or a file}
\usage{
write_chunk_from_array(v)
}
\arguments{
\item{v}{four-dimensional array with dimensions band, time, y, and x}
}
\description{
This function can be used within function passed to \code{\link{chunk_apply}} in order to pass four-dimensional R arrays as a
data cube chunk to the gdalcubes C++ library. It works only for R processes, which have been started from the gdalcubes C++ library. 
The input array must have dimensions band, time, y, x (in this order).
}
\note{
Call this function ONLY from a function passed to \code{\link{chunk_apply}}.

This function only works in R sessions started from gdalcubes streaming.
}
