% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{write_ncdf}
\alias{write_ncdf}
\title{Materialize a data cube as a NetCDF file}
\usage{
write_ncdf(x, fname = tempfile(pattern = "gdalcubes", fileext = ".nc"))
}
\arguments{
\item{x}{a data cube proxy object (class cube)}

\item{fname}{output file name}
}
\description{
This function will read chunks of a data cube and write them to a single NetCDF file.
}
\details{
The resulting NetCDF file contains three dimensions (t, y, x) and bands as variables.
}
