% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pixel.R
\name{apply_pixel}
\alias{apply_pixel}
\title{Apply a function over (multiband) pixels}
\usage{
apply_pixel(x, ...)
}
\arguments{
\item{object}{to be reduced}
}
\value{
return value and type depend on the class of x
}
\description{
This generic function applies a  function on pixels of a data cube, an R array, or other classes if implemented.
}
