% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as_stars}
\alias{as_stars}
\title{Coerce gdalcubes object into a stars object}
\usage{
as_stars(from)
}
\arguments{
\item{from}{data cube object to coerce}
}
\value{
stars object
}
\description{
The function materializes a data cube as a temporary NetCDF file and loads the file 
with the stars package.
}
