% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_collection.R
\name{gcbs_image_collection}
\alias{gcbs_image_collection}
\title{Load an existing image collection from a file}
\usage{
gcbs_image_collection(path)
}
\arguments{
\item{path}{Path to an existing image collection file}
}
\value{
An image collection proxy object, which can be used to create a data cube using \code{gcbs_cube}
}
\description{
This function will load an image collection from an SQLite file. These files
index / reference existing imagery. To create a collection from files on disk,
use \code{gcbs_create_collection}.
}
\examples{
\dontrun{gcbs_image_collection("/home/marius/github/gdalcubes/cmake-build-debug/src/test.db")}
}
