% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{gcbs_reduce}
\alias{gcbs_reduce}
\title{Reduce a data cube over the time dimension}
\usage{
gcbs_reduce(cube, reducer = c("mean", "median", "min", "max"))
}
\arguments{
\item{cube}{Source data cube}

\item{reducer}{Reducer function, currently "min", "max", "median", "mean", "count", "sd", "var", or "sum"}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which applies a reducer function over pixel time series of a data cube
}
\note{
Implemented reducers will ignore any NAN values (as na.rm=TRUE does)

This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.
}
\examples{
\dontrun{
gcbs_reduce(XXXX, "min")}
}
