% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bands.R
\name{gcbs_select_bands}
\alias{gcbs_select_bands}
\title{Select bands of a data cube}
\usage{
gcbs_select_bands(cube, bands)
}
\arguments{
\item{cube}{Input data cube}

\item{bands}{character vector with band names}
}
\value{
A proxy data cube object
}
\description{
Create a proxy data cube, which selects specific bands of a data cube. The resulting cube
will drop any other bands.
}
\note{
This function returns a proxy object, i.e., it will not start any computations besides deriving the shape of the result.

For performance reasons, gcbs_select_bands should always be called directly on a image collection cube created with gcbs_cube and 
drop all unneded bands.This allows to reduce GDAL RasterIO and warp operations.
}
\examples{
 L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                        ".TIF", recursive = TRUE, full.names = TRUE)
 v = gcbs_view(l=388941.2, r=766552.4, b=4345299, t=4744931, 
         proj="+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs",
         nx = 497, ny=526, t0="2018-01", t1="2018-12", dt="P1M")
 L8.col = gcbs_create_image_collection(L8_files, "L8_L1TP") 
 L8.cube = gcbs_cube(L8.col, v) 
 L8.rgb = gcbs_select_bands(L8.cube, c("B02", "B03", "B04"))
 L8.rgb
 L8.rgb.median = gcbs_reduce_time(L8.rgb, "median(B02)", "median(B03)", "median(B04)")  
 plot(L8.rgb.median, rgb=3:1, zlim=c(4000,12000))
}
