% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes.R
\docType{package}
\name{gdalcubes}
\alias{gdalcubes}
\alias{gdalcubes-package}
\title{gdalcubes: Earth observation data cubes from GDAL image collections}
\description{
The gdalcubes package constructs Earth Observation data cubes from many GDAL images / datasets..
}
