% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes.R
\docType{package}
\name{gdalcubes}
\alias{gdalcubes}
\alias{gdalcubes-package}
\title{gdalcubes: Streaming GDAL Data Cubes to R}
\description{
The gdalcubes package constructs Earth Observation data cubes from many GDAL images / datasets and allows streaming data chunk-wise into R.
}
