% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gdalcubes_set_ncdf_compression}
\alias{gdalcubes_set_ncdf_compression}
\title{Set compression level for netCDF files produced by gdalcubes}
\usage{
gdalcubes_set_ncdf_compression(level = 2)
}
\arguments{
\item{level}{integer; compression level, 0 = no compression, 1=fast compression, 9=small compression}
}
\description{
Set compression level for netCDF files produced by gdalcubes
}
\note{
THIS FUNCTION IS DEPRECATED AND IS GOING TO BE REPLACED BY \code{\link{gdalcubes_options}}.
}
\examples{
gdalcubes_set_ncdf_compression(9) # maximum compression
gdalcubes_set_ncdf_compression(0) # no compression
}
