% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{memsize}
\alias{memsize}
\title{Query data cube properties}
\usage{
memsize(obj, unit = "MiB")
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}

\item{unit}{Unit of data size, can be "B", "KB", "KiB", "MB", "MiB", "GB", "GiB", "TB", "TiB", "PB", "PiB"}
}
\value{
Total data size of data cube values expressed in the given unit
}
\description{
Query data cube properties
}
\examples{
# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db")) 
}

L8.col = image_collection(file.path(tempdir(), "L8.db"))
v = cube_view(extent=list(left=388941.2, right=766552.4, 
              bottom=4345299, top=4744931, t0="2018-04", t1="2018-06"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")
memsize(raster_cube(L8.col, v))
}
