% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cube}
\alias{predict.cube}
\title{Predict values based on a trained model to all pixels of a data cube.}
\usage{
\method{predict}{cube}(object, model, ...)
}
\arguments{
\item{object}{a data cube proxy object (class cube)}

\item{model}{model used for prediction (e.g. from \code{caret})}

\item{...}{further arguments passed to predict}
}
\description{
Predict values based on a trained model to all pixels of a data cube.
}
\details{
The predict method used will be automatically chosen based on the model class. It should support models from the \code{caret} 
package and simple models as from \code{lm} or \code{glm} . However, it  currently works only for a single output variable and 
requires a \code{data.frame} input.
}
\examples{
\donttest{
# create image collection from example Landsat data only
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db"), quiet = TRUE)
}

v = cube_view(extent=list(left=388941.2, right=766552.4,
                          bottom=4345299, top=4744931, t0="2018-04", t1="2018-06"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P3M")

L8.col = image_collection(file.path(tempdir(), "L8.db"))

x = sf::st_read(system.file("ny_samples.gpkg", package = "gdalcubes"))

raster_cube(L8.col, v) |>
  select_bands(c("B02","B03","B04","B05")) |>
  extract_geom(x) -> train

x$FID = rownames(x)  
train = merge(train, x, by = "FID")
train$iswater = as.factor(train$class == "water")

log_model <- glm(iswater ~ B02 + B03 + B04 + B05, data = train, family = "binomial")

raster_cube(L8.col, v) |>
  select_bands(c("B02","B03","B04","B05")) |>
  predict(model=log_model, type="response") |>
  plot(key.pos=1)
}
}
