% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{proj4}
\alias{proj4}
\title{Query data cube properties}
\usage{
proj4(obj)
}
\arguments{
\item{obj}{a data cube proxy object (class cube)}
}
\value{
The spatial reference system expressed as proj4 string
}
\description{
Query data cube properties
}
