% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{reduce_space}
\alias{reduce_space}
\title{Apply a function over space and bands in a four-dimensional (band, time, y, x) array}
\usage{
reduce_space(x, FUN, ...)
}
\arguments{
\item{x}{four-dimensional input bands with dimension order band, time, y, x}

\item{FUN}{function which receives one spatial slice in a three-dimensional array with dimensions bands, y, x as input}

\item{...}{further arguments passed to FUN}
}
\description{
Apply a function over space and bands in a four-dimensional (band, time, y, x) array
}
\details{
FUN is expected to produce a numeric vector (or scalar) where elements are interpreted as new bands in the result
}
\note{
This is a helper function that uses the same dimension ordering as gdalcubes streaming. It can be used to simplify 
the application of R functions e.g. over spatial slices in a data cube.
}
\examples{
load(system.file("extdata","sample_chunk.Rdata", package="gdalcubes"))
y = reduce_space(sample_chunk, function(x) {
 ndvi <- (x[8,,]-x[4,,])/(x[8,,]+x[4,,])
 return(c(min(ndvi, na.rm=TRUE),max(ndvi, na.rm=T)))
})
}
