% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming.R
\name{reduce_time}
\alias{reduce_time}
\title{Apply a function over time and bands in a four-dimensional (band, time, y, x) array}
\usage{
reduce_time(x, FUN, ...)
}
\arguments{
\item{x}{four-dimensional input bands with dimension order band, time, y, x}

\item{FUN}{function which receives one time series in a two-dimensional array with dimensions bands, time as input}

\item{...}{further arguments passed to FUN}
}
\description{
Apply a function over time and bands in a four-dimensional (band, time, y, x) array
}
\details{
FUN is expected to produce a numeric vector (or scalar) where elements are interpreted as new bands in the result
}
\note{
This is a helper function that uses the same dimension ordering as gdalcubes streaming. It can be used to simplify 
the application of R functions e.g. over time series in a data cube.
}
\examples{
\dontrun{
load(system.file("extdata","sample_chunk.Rdata", package="gdalcubes"))
reduce_time(sample_chunk, function(x) {
  ndvi <- (x[8,]-x[4,])/(x[8,]+x[4,])
  return(c(min(ndvi, na.rm=TRUE),max(ndvi, na.rm=T)))
})}
}
